+++ src/zabbix_enums/__init__.py
@@ -1,0 +1,8 @@
+from enum import IntEnum
+
+
+class _ZabbixEnum(IntEnum):
+    @classmethod
+    def _missing_(cls, value):
+        if isinstance(value, str):
+            return cls(int(value))
