--- src/zabbix_enums/common/discovered_host.py
@@ -1,6 +1,0 @@
-from zabbix_enums.common import _ZabbixEnum
-
-
-class DiscoveredHostStatus(_ZabbixEnum):
-    UP = 0
-    DOWN = 1
