--- src/zabbix_enums/common/discovered_service.py
@@ -1,6 +1,0 @@
-from zabbix_enums.common import _ZabbixEnum
-
-
-class DiscoveredServiceStatus(_ZabbixEnum):
-    UP = 0
-    DOWN = 1
