--- src/zabbix_enums/common/discovery_rule.py
@@ -1,6 +1,0 @@
-from zabbix_enums.common import _ZabbixEnum
-
-
-class DiscoveryRuleStatus(_ZabbixEnum):
-    ENABLED = 0
-    DISABLED = 1
