--- src/zabbix_enums/common/image.py
@@ -1,6 +1,0 @@
-from zabbix_enums.common import _ZabbixEnum
-
-
-class Image(_ZabbixEnum):
-    ICON = 1
-    BACKGROUND_IMAGE = 2
